# swagger-docs.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/modules": {
      "post": {
        "operationId": "ModulesController_createModule",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateModuleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Module created successfully"
          },
          "400": {
            "description": "Bad request"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new module",
        "tags": ["Modules"]
      },
      "get": {
        "operationId": "ModulesController_findAllModules",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (default: 1)",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of items per page (default: 10, max: 100)",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 10,
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "text",
            "required": false,
            "in": "query",
            "description": "Search text in title and description (case-insensitive)",
            "schema": {
              "example": "psychology",
              "type": "string"
            }
          },
          {
            "name": "difficulty",
            "required": false,
            "in": "query",
            "description": "Filter by difficulty level",
            "schema": {
              "enum": ["BEGINNER", "INTERMEDIATE", "ADVANCED"],
              "type": "string"
            }
          },
          {
            "name": "published",
            "required": false,
            "in": "query",
            "description": "Filter by published status (only available for non-student users)",
            "schema": {
              "example": true,
              "type": "boolean"
            }
          },
          {
            "name": "progress_status",
            "required": false,
            "in": "query",
            "description": "Filter by progress status (only available for students). Use ALL to show all modules regardless of progress status.",
            "schema": {
              "example": "IN_PROGRESS",
              "type": "string",
              "enum": ["NOT_STARTED", "IN_PROGRESS", "COMPLETED", "ALL"]
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Sort by field. progress_status is only available for students (ASC: IN_PROGRESS → NOT_STARTED → COMPLETED, DESC: COMPLETED → NOT_STARTED → IN_PROGRESS)",
            "schema": {
              "enum": [
                "title",
                "difficulty",
                "duration",
                "created_at",
                "progress_status"
              ],
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "description": "Sort order",
            "schema": {
              "enum": ["asc", "desc"],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Modules retrieved successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all modules with filtering and sorting",
        "tags": ["Modules"]
      }
    },
    "/api/modules/toggle-visibility": {
      "post": {
        "operationId": "ModulesController_toggleModuleVisibility",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ToggleModuleVisibilityDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Module visibility toggled successfully"
          },
          "400": {
            "description": "Bad request - Module must have at least one chapter and one quiz group to be published"
          },
          "404": {
            "description": "Module not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Toggle module visibility (publish/unpublish)",
        "tags": ["Modules"]
      }
    },
    "/api/modules/assign-professors": {
      "post": {
        "operationId": "ModulesController_assignProfessorsToModule",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignProfessorDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Professors assigned successfully"
          },
          "404": {
            "description": "Module or professor not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Assign professors to a module",
        "tags": ["Modules"]
      }
    },
    "/api/modules/unassign-professor": {
      "post": {
        "operationId": "ModulesController_unassignProfessorFromModule",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UnassignProfessorDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Professor unassigned successfully"
          },
          "404": {
            "description": "Module or professor not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Unassign a professor from a module",
        "tags": ["Modules"]
      }
    },
    "/api/modules/{id}/assignments": {
      "get": {
        "operationId": "ModulesController_getModuleAssignments",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Module ID",
            "schema": {
              "example": "507f1f77bcf86cd799439011",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Module assignments retrieved successfully"
          },
          "404": {
            "description": "Module not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all professors assigned to a module",
        "tags": ["Modules"]
      }
    },
    "/api/modules/assignments/audit-logs": {
      "get": {
        "operationId": "ModulesController_getAssignmentAuditLogs",
        "parameters": [
          {
            "name": "module_id",
            "required": false,
            "in": "query",
            "description": "Filter by module ID",
            "schema": {
              "example": "507f1f77bcf86cd799439011",
              "type": "string"
            }
          },
          {
            "name": "professor_id",
            "required": false,
            "in": "query",
            "description": "Filter by professor ID",
            "schema": {
              "example": "507f1f77bcf86cd799439011",
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (starts from 1)",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of items per page",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 10,
              "example": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Audit logs retrieved successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get assignment audit logs for admin accountability",
        "tags": ["Modules"]
      }
    },
    "/api/modules/{id}": {
      "get": {
        "operationId": "ModulesController_findModuleById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Module ID",
            "schema": {
              "example": "507f1f77bcf86cd799439011",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Module retrieved successfully"
          },
          "404": {
            "description": "Module not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a module by id",
        "tags": ["Modules"]
      },
      "patch": {
        "operationId": "ModulesController_updateModule",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Module ID",
            "schema": {
              "example": "507f1f77bcf86cd799439011",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateModuleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Module updated successfully"
          },
          "404": {
            "description": "Module not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a module",
        "tags": ["Modules"]
      },
      "delete": {
        "operationId": "ModulesController_removeModule",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Module ID",
            "schema": {
              "example": "507f1f77bcf86cd799439011",
              "type": "string"
            }
          },
          {
            "name": "school_id",
            "required": false,
            "in": "query",
            "description": "School ID (required for super admin, optional for other roles)",
            "schema": {
              "example": "507f1f77bcf86cd799439011",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Module deleted successfully"
          },
          "404": {
            "description": "Module not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a module",
        "tags": ["Modules"]
      }
    },
...
  },
  "info": {
    "title": "AI Professor API",
    "description": "API documentation for AI Professor backend",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateModuleDto": {
        "type": "object",
        "properties": {
          "school_id": {
            "type": "object",
            "example": "507f1f77bcf86cd799439011",
            "description": "School ID (required for super admin, optional for other roles)"
          },
          "title": {
            "type": "string",
            "example": "Child and Adolescent Development Psychology",
            "description": "Title of the module"
          },
          "subject": {
            "type": "string",
            "example": "Child Development",
            "description": "Subject of the module"
          },
          "description": {
            "type": "string",
            "example": "This module covers the fundamental principles of child and adolescent development psychology, including cognitive, social, and emotional development stages.",
            "description": "Overview/intro shown in UI"
          },
          "category": {
            "type": "string",
            "example": "Psychology",
            "description": "Category of the module"
          },
          "difficulty": {
            "type": "string",
            "example": "INTERMEDIATE",
            "enum": ["BEGINNER", "INTERMEDIATE", "ADVANCED"],
            "description": "Difficulty level of the module"
          },
          "tags": {
            "example": [
              "psychology",
              "child-development",
              "cognitive-development"
            ],
            "description": "Array of tags for better searching and filtering",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "thumbnail": {
            "type": "string",
            "example": "/images/module-thumbnails/child-development.jpg",
            "description": "Thumbnail image path for the module"
          }
        },
        "required": ["title", "subject", "description", "difficulty"]
      },
      "ToggleModuleVisibilityDto": {
        "type": "object",
        "properties": {
          "school_id": {
            "type": "object",
            "example": "507f1f77bcf86cd799439011",
            "description": "School ID (required for super admin, optional for other roles)"
          },
          "module_id": {
            "type": "object",
            "example": "507f1f77bcf86cd799439011",
            "description": "ID of the module to toggle visibility"
          },
          "action": {
            "type": "string",
            "enum": ["PUBLISH", "UNPUBLISH"],
            "example": "PUBLISH",
            "description": "Action to perform: PUBLISH or UNPUBLISH"
          }
        },
        "required": ["module_id", "action"]
      },
      "ProfessorAssignmentDto": {
        "type": "object",
        "properties": {
          "professor_id": {
            "type": "object",
            "example": "507f1f77bcf86cd799439011",
            "description": "Professor ID"
          }
        },
        "required": ["professor_id"]
      },
      "AssignProfessorDto": {
        "type": "object",
        "properties": {
          "module_id": {
            "type": "object",
            "example": "507f1f77bcf86cd799439011",
            "description": "Module ID"
          },
          "assignments": {
            "description": "Array of professor assignments",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProfessorAssignmentDto"
            }
          }
        },
        "required": ["module_id", "assignments"]
      },
      "UnassignProfessorDto": {
        "type": "object",
        "properties": {
          "module_id": {
            "type": "object",
            "example": "507f1f77bcf86cd799439011",
            "description": "Module ID"
          },
          "professor_id": {
            "type": "object",
            "example": "507f1f77bcf86cd799439011",
            "description": "Professor ID"
          }
        },
        "required": ["module_id", "professor_id"]
      },
      "UpdateModuleDto": {
        "type": "object",
        "properties": {
          "school_id": {
            "type": "object",
            "example": "507f1f77bcf86cd799439011",
            "description": "School ID (required for super admin, optional for other roles)"
          },
          "title": {
            "type": "string",
            "example": "Child and Adolescent Development Psychology",
            "description": "Title of the module"
          },
          "subject": {
            "type": "string",
            "example": "Child Development",
            "description": "Subject of the module"
          },
          "description": {
            "type": "string",
            "example": "This module covers the fundamental principles of child and adolescent development psychology, including cognitive, social, and emotional development stages.",
            "description": "Overview/intro shown in UI"
          },
          "category": {
            "type": "string",
            "example": "Psychology",
            "description": "Category of the module"
          },
          "difficulty": {
            "type": "string",
            "example": "INTERMEDIATE",
            "enum": ["BEGINNER", "INTERMEDIATE", "ADVANCED"],
            "description": "Difficulty level of the module"
          },
          "tags": {
            "example": [
              "psychology",
              "child-development",
              "cognitive-development"
            ],
            "description": "Array of tags for better searching and filtering",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "thumbnail": {
            "type": "string",
            "example": "/images/module-thumbnails/child-development.jpg",
            "description": "Thumbnail image path for the module"
          }
        }
      },
...
    }
  }
}